[Stability/Perf] Remove extra completers
It is less than ideal having completers and their dependencies merged
into this repository.

Initially, this work will only concerned with Swift and distilling YCMD
down to the essential features.

- Most completers other than swift are gone
- Remove special cased clang code
- Fix shutdown test
- Rip out things from third party
- Some readme cleanup

In the future, there should not be any completer specific code in here
but swift is left now for simplicity